chore(#346): rename java package structure for gateway
