+++ digiwf-gateway/src/main/java/de/muenchen/oss/digiwf/gateway/configuration/NoSecurityConfiguration.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
+ * der Landeshauptstadt München, 2022
+ */
+package de.muenchen.oss.digiwf.gateway.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.security.config.web.server.ServerHttpSecurity;
+import org.springframework.security.web.server.SecurityWebFilterChain;
+
+
+@Configuration
+@Profile("no-security")
+public class NoSecurityConfiguration {
+
+    @Bean
+    public SecurityWebFilterChain springSecurityFilterChain(ServerHttpSecurity http) {
+        return http
+                .authorizeExchange()
+                        .anyExchange().permitAll()
+                .and()
+                    .cors()
+                .and()
+                    .csrf().disable()
+                .build();
+    }
+
+}
