--- digiwf-gateway/src/main/java/io/muenchendigital/digiwf/gateway/exception/ParameterPollutionException.java
@@ -1,24 +1,0 @@
-/*
- * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
- * der Landeshauptstadt München, 2022
- */
-package io.muenchendigital.digiwf.gateway.exception;
-
-import io.muenchendigital.digiwf.gateway.filter.GlobalRequestParameterPollutionFilter;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-
-/**
- * Used in {@link GlobalRequestParameterPollutionFilter} to signal a possible parameter pollution attack.
- */
-@ResponseStatus(code = HttpStatus.BAD_REQUEST, reason = "parameter pollution")
-@NoArgsConstructor
-@EqualsAndHashCode(callSuper = true)
-@ToString(callSuper = true)
-public class ParameterPollutionException extends RuntimeException {
-    // default Ctor
-}
