+++ digiwf-gateway/src/test/java/de/muenchen/oss/digiwf/gateway/TestConstants.java
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
+ * der Landeshauptstadt München, 2022
+ */
+package de.muenchen.oss.digiwf.gateway;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class TestConstants {
+
+    public static final String SPRING_TEST_PROFILE = "test";
+
+}
