--- digiwf-gateway/src/test/java/io/muenchendigital/digiwf/gateway/TestConstants.java
@@ -1,16 +1,0 @@
-/*
- * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
- * der Landeshauptstadt München, 2022
- */
-package io.muenchendigital.digiwf.gateway;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class TestConstants {
-
-    public static final String SPRING_TEST_PROFILE = "test";
-    
-}
