--- digiwf-gateway/src/test/java/io/muenchendigital/digiwf/gateway/filter/CsrfTokenAppendingHelperFilterTest.java
@@ -1,39 +1,0 @@
-/*
- * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
- * der Landeshauptstadt München, 2022
- */
-package io.muenchendigital.digiwf.gateway.filter;
-
-import io.muenchendigital.digiwf.gateway.ApiGatewayApplication;
-import io.muenchendigital.digiwf.gateway.TestConstants;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.web.reactive.server.WebTestClient;
-
-
-@ExtendWith(SpringExtension.class)
-@SpringBootTest(
-        classes = { ApiGatewayApplication.class },
-        webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT
-)
-@ActiveProfiles(TestConstants.SPRING_TEST_PROFILE)
-
-class CsrfTokenAppendingHelperFilterTest {
-
-    @Autowired
-    private WebTestClient webTestClient;
-
-    @Test
-    @WithMockUser
-    void csrfCookieAppendition() {
-        webTestClient.get().uri("/").exchange()
-                .expectHeader()
-                    .valueMatches("set-cookie", "XSRF-TOKEN=[a-f\\d]{8}(-[a-f\\d]{4}){3}-[a-f\\d]{12}?;\\sPath=/");
-    }
-
-}
