--- digiwf-gateway/src/test/java/io/muenchendigital/digiwf/gateway/filter/GlobalRequestParameterPollutionFilterTest.java
@@ -1,47 +1,0 @@
-/*
- * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
- * der Landeshauptstadt München, 2022
- */
-package io.muenchendigital.digiwf.gateway.filter;
-
-import io.muenchendigital.digiwf.gateway.ApiGatewayApplication;
-import io.muenchendigital.digiwf.gateway.TestConstants;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.web.reactive.server.WebTestClient;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import java.nio.charset.StandardCharsets;
-
-
-@ExtendWith(SpringExtension.class)
-@SpringBootTest(
-        classes = { ApiGatewayApplication.class },
-        webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT
-)
-@ActiveProfiles(TestConstants.SPRING_TEST_PROFILE)
-class GlobalRequestParameterPollutionFilterTest {
-
-    @Autowired
-    private WebTestClient webTestClient;
-
-    @Test
-    @WithMockUser
-    void parameterPollutionAttack() {
-        final StringBuilder jsonResponseBody = new StringBuilder();
-        webTestClient.get().uri("/api/frontend-backend-service/testendpoint?parameter1=testdata_1&parameter2=testdata&parameter1=testdata_2").exchange()
-                .expectStatus()
-                    .isEqualTo(HttpStatus.BAD_REQUEST)
-                .expectBody()
-                    .consumeWith(responseBody -> jsonResponseBody.append(new String(responseBody.getResponseBody(), StandardCharsets.UTF_8)));
-        assertTrue(jsonResponseBody.toString().contains("\"message\" : \"parameter pollution\""));
-    }
-
-}
